Extract libs/policyengine-simulation-observability
Move the observability plumbing shared by the gateway and executor —
the policyengine-observability configuration wrapper, legacy Logfire
helpers, error redaction, and the telemetry envelope — into a lib with
its own pyproject and uv.lock. Isolated from the upcoming contract lib
because it changes for a different reason: the Logfire pieces are
retained only while a replacement platform is evaluated, and this
dependency cluster is the one that caused #602. Both Modal images now
mount the lib alongside the project sources, with the mount tuple
pinned by a new image-structure assertion.
