+++ libs/policyengine-simulation-observability/src/policyengine_simulation_observability/__init__.py
@@ -1,0 +1,1 @@
+"""Shared observability plumbing for the simulation gateway and executor."""
