+++ projects/policyengine-simulation-executor/pyproject.toml
@@ -16,6 +16,7 @@ dependencies = [
     "pydantic-settings (>=2.7.1,<3.0.0)",
     "opentelemetry-instrumentation-fastapi (>=0.51b0,<0.52)",
     "policyengine-fastapi",
+    "policyengine-simulation-observability",
     "policyengine==4.18.9",
     "policyengine-core==3.28.0",
     "policyengine-uk==2.89.2",
@@ -60,6 +61,7 @@ modal-gateway-image = [
 
 [tool.uv.sources]
 policyengine-fastapi = { path = "../../libs/policyengine-fastapi", editable = true }
+policyengine-simulation-observability = { path = "../../libs/policyengine-simulation-observability", editable = true }
 
 [project.optional-dependencies]
 test = [ "pytest>=8.3.4", "pytest-asyncio>=0.25.3", "pytest-cov>=6.1.1",]
