+++ projects/policyengine-simulation-executor/src/policyengine_simulation_executor/main.py
@@ -2,7 +2,7 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI
 from policyengine_fastapi.exit import exit
 from policyengine_simulation_executor import initialize
-from policyengine_simulation_executor.observability import (
+from policyengine_simulation_observability.observability import (
     init_simulation_observability,
 )
 from policyengine_fastapi import ping
