+++ projects/policyengine-simulation-executor/src/policyengine_simulation_executor/simulation_output_builder.py
@@ -14,7 +14,7 @@ from policyengine_simulation_executor import simulation_output_geographic
 from policyengine_simulation_executor import simulation_output_inequality
 from policyengine_simulation_executor import simulation_output_labor
 from policyengine_simulation_executor import simulation_output_poverty
-from policyengine_simulation_executor.observability import SegmentName
+from policyengine_simulation_observability.observability import SegmentName
 from policyengine_simulation_executor.release_bundle import get_country_release_bundle
 from policyengine_simulation_executor.simulation_macro_output import (
     BudgetaryImpact,
