+++ projects/policyengine-simulation-executor/tests/test_simulation_output_builder.py
@@ -30,7 +30,7 @@ from fixtures.test_simulation_output_builder import (
 )
 from policyengine_simulation_executor.release_bundle import BUNDLE_RECEIPT_FILENAME
 from policyengine_simulation_executor.release_bundle import get_country_release_bundle
-from policyengine_simulation_executor.observability import SegmentName
+from policyengine_simulation_observability.observability import SegmentName
 from policyengine_simulation_executor.simulation_runtime import RegionResolution
 from policyengine_simulation_executor.simulation_runtime import _load_dataset
 from policyengine_simulation_executor.simulation_runtime import _normalise_policy
